+++ include/Pixel.h
@@ -13,8 +13,8 @@ public:
 	// Get colour of the pixel
 	ColorDBL getColor() const;
 
-	// Set the colour of the pixel, with double values
-	void setColor(double _r, double _g, double _b);
+	// Set the colour of the pixel, with float values
+	void setColor(float _r, float _g, float _b);
 
 	// Set the colour of the pixel, with ColorDBL
 	void setColor(ColorDBL _color);
